feat(jstzd): move files
