--- crates/jstzd/src/task/task.rs
@@ -1,20 +1,0 @@
-use super::joinhandle::JoinHandle;
-use async_trait::async_trait;
-use tokio::signal::unix::Signal;
-use tokio::sync::mpsc;
-
-#[async_trait]
-pub trait Task {
-    /// A set of tasks which are all executed on the same 'execution backend'
-    type TaskSet;
-
-    /// Performs the unit of work for the task
-    async fn run_task(
-        self,
-        signals: mpsc::Receiver<Signal>,
-        task_group: &Self::TaskSet,
-    ) -> anyhow::Result<()>;
-
-    /// Spawns the task
-    async fn spawn_task(self) -> JoinHandle;
-}
